+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [1.3.2] (Unreleased)
+
+* **Fix**: [142](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/issues/142) Added
+  field to provide base64 string data for profile picture.
 * **Fix**: [126](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/issues/126) Added
   flag to hide user name in chat.
 * **Feat**: [161](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/161) Added